ASC-918 Refactor Integration Test Suite
A heavy lift in the integration test suite fixtures was required in order to
properly validate test steps. Separate commit solely to ease PR review
process.
