+++ tests/integration/test_negative.py
@@ -5,9 +5,10 @@
 # ======================================================================================================================
 # Imports
 # ======================================================================================================================
+from __future__ import absolute_import
 import pytest
 import swagger_client
-from .conftest import ZigZagRunner
+from tests.helper.classes.zigzag_runner import ZigZagRunner
 
 
 # ======================================================================================================================
